Fix a CSS bug in recent versions of WebKit and Gecko
Between firefox versions 130 and 134, something broke vertical
alignment. Interestingly, the exact same issue can be observed in
WebKit, and introduced roughly at the same time, which suggests that
it might be a deliberate change, and not a bug?

This commit adds a trivial padding change to move things back into
place. I'll try to find a way to make a repro case and open a bug, to
figure out exactly *why* this broke after almost a decade.